Fix a few Coverity static analysis bugs.

+++ src/glChartCanvas.cpp
@@ -3621,14 +3621,14 @@ void glChartCanvas::RenderWorldChart(ocpnDC &dc, ViewPort &vp, wxRect &rect, boo
             tvp.clat = 0, tvp.clon = 0;
             tvp.rotation = 0;
             wxPoint2DDouble p = tvp.GetDoublePixFromLL( 89.99, 0);
-            float w = tvp.pix_width/2, h = tvp.pix_height/2;
+            float w = ((float)tvp.pix_width)/2, h = ((float)tvp.pix_height)/2;
             double world_r = h - p.m_y;
             const float pi_ovr100 = float(M_PI)/100;
             if(world_r*world_r < w*w + h*h) {
                 glClear( GL_COLOR_BUFFER_BIT );
 
                 glBegin(GL_TRIANGLE_FAN);
-                float w = vp.pix_width/2, h = vp.pix_height/2;
+                float w = ((float)vp.pix_width)/2, h = ((float)vp.pix_height)/2;
                 for(float theta = 0; theta < 2*M_PI+.01f; theta+=pi_ovr100)
                     glVertex2f(w + world_r*sinf(theta), h + world_r*cosf(theta));
                 glEnd();
